+++ src/Type/TurnstileType.php
@@ -11,22 +11,15 @@ use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+/**
+ * @extends AbstractType<never>
+ */
 class TurnstileType extends AbstractType
 {
-    /**
-     * @var string
-     */
-    private $key;
-
-    /**
-     * @var bool
-     */
-    private $enable;
-
-    public function __construct(string $key, bool $enable)
-    {
-        $this->key = $key;
-        $this->enable = $enable;
+    public function __construct(
+        private readonly string $key,
+        private readonly bool $enable,
+    ) {
     }
 
     public function configureOptions(OptionsResolver $resolver): void